+++ src/plugins/nut/models/group.rs
@@ -11,28 +11,28 @@ use super::{
 #[derive(Queryable, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Item {
-    pub id: i32,
+    pub id: i64,
     pub code: String,
     pub name: String,
-    pub parent_id: Option<i32>,
-    pub level: i32,
-    pub version: i32,
+    pub parent_id: Option<i64>,
+    pub level: i64,
+    pub version: i64,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
 }
 
 pub trait Dao {
     fn all(&self) -> Result<Vec<Item>>;
-    fn by_id(&self, id: i32) -> Result<Item>;
+    fn by_id(&self, id: i64) -> Result<Item>;
     fn by_code(&self, cod: &str) -> Result<Item>;
-    fn by_user(&self, user: i32) -> Result<Vec<Item>>;
-    fn offsprings(&self, id: i32) -> Result<Vec<Item>>;
-    fn users(&self, id: i32) -> Result<Vec<User>>;
-    fn create(&self, code: &str, name: &str, parent_id: Option<i32>) -> Result<()>;
-    fn update(&self, id: i32, code: &str, name: &str) -> Result<()>;
-    fn associate(&self, group: i32, user: i32) -> Result<()>;
-    fn dissociate(&self, group: i32, user: i32) -> Result<()>;
-    fn destory(&self, id: i32) -> Result<()>;
+    fn by_user(&self, user: i64) -> Result<Vec<Item>>;
+    fn offsprings(&self, id: i64) -> Result<Vec<Item>>;
+    fn users(&self, id: i64) -> Result<Vec<User>>;
+    fn create(&self, code: &str, name: &str, parent_id: Option<i64>) -> Result<()>;
+    fn update(&self, id: i64, code: &str, name: &str) -> Result<()>;
+    fn associate(&self, group: i64, user: i64) -> Result<()>;
+    fn dissociate(&self, group: i64, user: i64) -> Result<()>;
+    fn destory(&self, id: i64) -> Result<()>;
 }
 
 impl Dao for Connection {
@@ -42,7 +42,7 @@ impl Dao for Connection {
             .load(self)?;
         Ok(items)
     }
-    fn by_id(&self, id: i32) -> Result<Item> {
+    fn by_id(&self, id: i64) -> Result<Item> {
         let it = groups::dsl::groups
             .filter(groups::dsl::id.eq(id))
             .first(self)?;
@@ -54,8 +54,8 @@ impl Dao for Connection {
             .first(self)?;
         Ok(it)
     }
-    fn by_user(&self, user: i32) -> Result<Vec<Item>> {
-        let ids: Vec<i32> = groups_users::dsl::groups_users
+    fn by_user(&self, user: i64) -> Result<Vec<Item>> {
+        let ids: Vec<i64> = groups_users::dsl::groups_users
             .select(groups_users::dsl::group_id)
             .filter(groups_users::dsl::user_id.eq(user))
             .order(groups_users::dsl::created_at.desc())
@@ -67,7 +67,7 @@ impl Dao for Connection {
         }
         Ok(items)
     }
-    fn offsprings(&self, id: i32) -> Result<Vec<Item>> {
+    fn offsprings(&self, id: i64) -> Result<Vec<Item>> {
         let mut children = Vec::new();
 
         for it in groups::dsl::groups
@@ -79,8 +79,8 @@ impl Dao for Connection {
         }
         Ok(children)
     }
-    fn users(&self, id: i32) -> Result<Vec<User>> {
-        let ids: Vec<i32> = groups_users::dsl::groups_users
+    fn users(&self, id: i64) -> Result<Vec<User>> {
+        let ids: Vec<i64> = groups_users::dsl::groups_users
             .select(groups_users::dsl::user_id)
             .filter(groups_users::dsl::group_id.eq(id))
             .order(groups_users::dsl::created_at.desc())
@@ -92,7 +92,7 @@ impl Dao for Connection {
         }
         Ok(items)
     }
-    fn create(&self, code: &str, name: &str, parent: Option<i32>) -> Result<()> {
+    fn create(&self, code: &str, name: &str, parent: Option<i64>) -> Result<()> {
         let now = Utc::now().naive_local();
         let level = match parent {
             Some(id) => {
@@ -112,7 +112,7 @@ impl Dao for Connection {
             .execute(self)?;
         Ok(())
     }
-    fn update(&self, id: i32, code: &str, name: &str) -> Result<()> {
+    fn update(&self, id: i64, code: &str, name: &str) -> Result<()> {
         let now = Utc::now().naive_local();
         update(groups::dsl::groups.filter(groups::dsl::id.eq(&id)))
             .set((
@@ -123,7 +123,7 @@ impl Dao for Connection {
             .execute(self)?;
         Ok(())
     }
-    fn associate(&self, group: i32, user: i32) -> Result<()> {
+    fn associate(&self, group: i64, user: i64) -> Result<()> {
         let c: i64 = groups_users::dsl::groups_users
             .filter(groups_users::dsl::group_id.eq(group))
             .filter(groups_users::dsl::user_id.eq(user))
@@ -139,7 +139,7 @@ impl Dao for Connection {
         }
         Ok(())
     }
-    fn dissociate(&self, group: i32, user: i32) -> Result<()> {
+    fn dissociate(&self, group: i64, user: i64) -> Result<()> {
         delete(
             groups_users::dsl::groups_users
                 .filter(groups_users::dsl::group_id.eq(group))
@@ -148,11 +148,11 @@ impl Dao for Connection {
         .execute(self)?;
         Ok(())
     }
-    fn destory(&self, id: i32) -> Result<()> {
+    fn destory(&self, id: i64) -> Result<()> {
         let now = Utc::now().naive_local();
         update(groups::dsl::groups.filter(groups::dsl::parent_id.eq(id)))
             .set((
-                groups::dsl::parent_id.eq(None::<i32>),
+                groups::dsl::parent_id.eq(None::<i64>),
                 groups::dsl::updated_at.eq(&now),
             ))
             .execute(self)?;
