+++ src/plugins/nut/models/resource.rs
@@ -8,11 +8,11 @@ use super::schema::{policies, resources};
 #[derive(Queryable, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Item {
-    pub id: i32,
+    pub id: i64,
     pub type_: String,
     pub code: String,
     pub name: String,
-    pub version: i32,
+    pub version: i64,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
 }
@@ -24,12 +24,12 @@ impl Item {
 
 pub trait Dao {
     fn all(&self) -> Result<Vec<Item>>;
-    fn by_id(&self, id: i32) -> Result<Item>;
+    fn by_id(&self, id: i64) -> Result<Item>;
     fn by_code(&self, user: &str) -> Result<Item>;
     fn by_type_and_code(&self, type_: &str, code: &str) -> Result<Item>;
     fn create(&self, type_: &str, code: &str, name: &str) -> Result<()>;
-    fn update(&self, id: i32, type_: &str, code: &str, name: &str) -> Result<()>;
-    fn destory(&self, id: i32) -> Result<()>;
+    fn update(&self, id: i64, type_: &str, code: &str, name: &str) -> Result<()>;
+    fn destory(&self, id: i64) -> Result<()>;
 }
 
 impl Dao for Connection {
@@ -39,7 +39,7 @@ impl Dao for Connection {
             .load(self)?;
         Ok(items)
     }
-    fn by_id(&self, id: i32) -> Result<Item> {
+    fn by_id(&self, id: i64) -> Result<Item> {
         let it = resources::dsl::resources
             .filter(resources::dsl::id.eq(id))
             .first(self)?;
@@ -70,7 +70,7 @@ impl Dao for Connection {
             .execute(self)?;
         Ok(())
     }
-    fn update(&self, id: i32, type_: &str, code: &str, name: &str) -> Result<()> {
+    fn update(&self, id: i64, type_: &str, code: &str, name: &str) -> Result<()> {
         let now = Utc::now().naive_local();
         update(resources::dsl::resources.filter(resources::dsl::id.eq(&id)))
             .set((
@@ -82,7 +82,7 @@ impl Dao for Connection {
             .execute(self)?;
         Ok(())
     }
-    fn destory(&self, id: i32) -> Result<()> {
+    fn destory(&self, id: i64) -> Result<()> {
         delete(policies::dsl::policies.filter(policies::dsl::resource_id.eq(id))).execute(self)?;
         delete(resources::dsl::resources.filter(resources::dsl::id.eq(id))).execute(self)?;
         Ok(())
