+++ src/plugins/nut/models/tag.rs
@@ -8,22 +8,22 @@ use super::schema::tags;
 #[derive(Queryable, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Item {
-    pub id: i32,
+    pub id: i64,
     pub code: String,
     pub name: String,
     pub color: String,
-    pub version: i32,
+    pub version: i64,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
 }
 
 pub trait Dao {
     fn all(&self) -> Result<Vec<Item>>;
-    fn by_id(&self, id: i32) -> Result<Item>;
+    fn by_id(&self, id: i64) -> Result<Item>;
     fn by_code(&self, cod: &str) -> Result<Item>;
     fn create(&self, code: &str, name: &str, color: &str) -> Result<()>;
-    fn update(&self, id: i32, code: &str, name: &str, color: &str) -> Result<()>;
-    fn destory(&self, id: i32) -> Result<()>;
+    fn update(&self, id: i64, code: &str, name: &str, color: &str) -> Result<()>;
+    fn destory(&self, id: i64) -> Result<()>;
 }
 
 impl Dao for Connection {
@@ -31,7 +31,7 @@ impl Dao for Connection {
         let items = tags::dsl::tags.order(tags::dsl::name.asc()).load(self)?;
         Ok(items)
     }
-    fn by_id(&self, id: i32) -> Result<Item> {
+    fn by_id(&self, id: i64) -> Result<Item> {
         let it = tags::dsl::tags.filter(tags::dsl::id.eq(id)).first(self)?;
         Ok(it)
     }
@@ -53,7 +53,7 @@ impl Dao for Connection {
             .execute(self)?;
         Ok(())
     }
-    fn update(&self, id: i32, code: &str, name: &str, color: &str) -> Result<()> {
+    fn update(&self, id: i64, code: &str, name: &str, color: &str) -> Result<()> {
         let now = Utc::now().naive_local();
         update(tags::dsl::tags.filter(tags::dsl::id.eq(&id)))
             .set((
@@ -65,7 +65,7 @@ impl Dao for Connection {
             .execute(self)?;
         Ok(())
     }
-    fn destory(&self, id: i32) -> Result<()> {
+    fn destory(&self, id: i64) -> Result<()> {
         delete(tags::dsl::tags.filter(tags::dsl::id.eq(id))).execute(self)?;
         Ok(())
     }
